Fix ufdCopy() for large (> 2GB) files
- Files can be (much) larger than INT32_MAX, change the return
  type to off_t and fix + simplify the calculations. Fixes the other
  half of RhBug:790396 and makes ufdCopy() usable for other purposes too.

+++ rpmio/rpmio.h
@@ -116,7 +116,7 @@ FD_t fdFree(FD_t fd);
 
 /**
  */
-int ufdCopy(FD_t sfd, FD_t tfd);
+off_t ufdCopy(FD_t sfd, FD_t tfd);
 
 /** \ingroup rpmio
  * Identify per-desciptor I/O operation statistics.
